collapse specific subject type controllers into a single Subjects_Controller class
