+++ src/plugin/class-subjects-controller.php
@@ -7,15 +7,311 @@ use DateTimeZone;
 use WP_Error;
 use WP_REST_Controller;
 use WP_REST_Response;
+use WP_REST_Server;
 
 class Subjects_Controller extends WP_REST_Controller {
+	private string $storage_post_type;
+	private array $subject_types;
 
-	protected string $storage_post_type;
+	protected $schema;
 
 	public function __construct( $storage_post_type ) {
 		$this->storage_post_type = $storage_post_type;
+		$this->init_schema();
+		$this->subject_types = array_keys( $this->schema );
+
+		add_action( 'rest_api_init', array( $this, 'register_routes' ) );
 
 		$this->attach_filters();
+		$this->get_item_schema();
+	}
+
+	public function init_schema(): void {
+		$this->schema = array(
+			'blog-post' => array(
+				'$schema'    => 'http://json-schema.org/draft-04/schema#',
+				'title'      => 'blogpost',
+				'type'       => 'object',
+				'properties' => array(
+					'id'            => array(
+						'description' => __( 'Unique identifier for liberated blogpost', 'try_wordpress' ),
+						'type'        => 'integer',
+						'context'     => array( 'view', 'edit' ),
+						'readonly'    => true,
+					),
+					'authorId'      => array(
+						'description' => __( 'Author ID of the blogpost', 'try_wordpress' ),
+						'type'        => 'integer',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+					),
+					'sourceUrl'     => array(
+						'description' => __( 'Source URL from where the blogpost was liberated', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => true,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'sourceHtml'    => array(
+						'description' => __( 'Source HTML from where the blogpost was liberated', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'rawTitle'      => array(
+						'description' => __( 'Raw title of the blogpost', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'parsedTitle'   => array(
+						'description' => __( 'Parsed title of the blogpost', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'rawDate'       => array(
+						'description' => __( 'Raw date of the blogpost', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'parsedDate'    => array(
+						'description' => __( 'Parsed date of the blogpost', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'rawContent'    => array(
+						'description' => __( 'Raw content of the blogpost', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'parsedContent' => array(
+						'description' => __( 'Parsed content of the blogpost', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'transformedId' => array(
+						'description' => __( 'Post ID of transformed result of this liberated blogpost', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+				),
+			),
+			'page'      => array(
+				'$schema'    => 'http://json-schema.org/draft-04/schema#',
+				'title'      => 'page',
+				'type'       => 'object',
+				'properties' => array(
+					'id'            => array(
+						'description' => __( 'Unique identifier for liberated page', 'try_wordpress' ),
+						'type'        => 'integer',
+						'context'     => array( 'view', 'edit' ),
+						'readonly'    => true,
+					),
+					'authorId'      => array(
+						'description' => __( 'Author ID of the page', 'try_wordpress' ),
+						'type'        => 'integer',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+					),
+					'sourceUrl'     => array(
+						'description' => __( 'Source URL from where the page was liberated', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => true,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'sourceHtml'    => array(
+						'description' => __( 'Source HTML from where the page was liberated', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'rawTitle'      => array(
+						'description' => __( 'Raw title of the page', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'parsedTitle'   => array(
+						'description' => __( 'Parsed title of the page', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'rawDate'       => array(
+						'description' => __( 'Raw date of the page', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'parsedDate'    => array(
+						'description' => __( 'Parsed date of the page', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'rawContent'    => array(
+						'description' => __( 'Raw content of the page', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'parsedContent' => array(
+						'description' => __( 'Parsed content of the page', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+					'transformedId' => array(
+						'description' => __( 'Post ID of transformed result of this liberated page', 'try_wordpress' ),
+						'type'        => 'string',
+						'context'     => array( 'view', 'edit' ),
+						'required'    => false,
+						'arg_options' => array(
+							'sanitize_callback' => 'sanitize_text_field',
+						),
+					),
+				),
+			),
+		);
+	}
+
+	public function register_routes(): void {
+		$version   = '1';
+		$namespace = 'try-wp/v' . $version;
+
+		foreach ( $this->subject_types as $subject_type ) {
+			register_rest_route(
+				$namespace,
+				'/subjects/' . $subject_type,
+				array(
+					array(
+						'methods'             => WP_REST_Server::CREATABLE,
+						'callback'            => array( $this, 'create_item' ),
+						'permission_callback' => '__return_true',
+						'args'                => $this->get_endpoint_args_for_subject_schema( $subject_type, WP_REST_Server::CREATABLE ),
+					),
+					array(
+						'methods'             => WP_REST_Server::READABLE,
+						'callback'            => array( $this, 'search_item' ),
+						'permission_callback' => '__return_true',
+						'args'                => $this->get_endpoint_args_for_subject_schema( $subject_type, WP_REST_Server::READABLE ),
+					),
+				)
+			);
+
+			register_rest_route(
+				$namespace,
+				'/subjects/' . $subject_type . '/(?P<id>\d+)',
+				array(
+					array(
+						'methods'             => WP_REST_Server::READABLE,
+						'callback'            => array( $this, 'get_item' ),
+						'permission_callback' => '__return_true',
+						'args'                => array(
+							'context' => array(
+								'default' => 'view',
+							),
+						),
+					),
+					array(
+						'methods'             => WP_REST_Server::EDITABLE,
+						'callback'            => array( $this, 'update_item' ),
+						'permission_callback' => '__return_true',
+						'args'                => $this->get_endpoint_args_for_subject_schema( $subject_type, WP_REST_Server::EDITABLE ),
+					),
+					array(
+						'methods'             => WP_REST_Server::DELETABLE,
+						'callback'            => array( $this, 'delete_item' ),
+						'permission_callback' => '__return_true',
+						'args'                => array(
+							'force' => array(
+								'default' => false,
+							),
+						),
+					),
+				)
+			);
+
+			register_rest_route(
+				$namespace,
+				'/subjects/' . $subject_type . '/schema',
+				array(
+					'methods'             => WP_REST_Server::READABLE,
+					'callback'            => function () use ( $subject_type ) {
+						return $this->get_public_subject_schema( $subject_type );
+					},
+					'permission_callback' => '__return_true',
+				)
+			);
+		}
+	}
+
+	public function get_public_subject_schema( $subject_type ) {
+		$schema = $this->schema[ $subject_type ];
+		return $this->remove_arg_options_from_schema( $schema );
+	}
+
+	private function remove_arg_options_from_schema( &$schema ) {
+		foreach ( $schema['properties'] as &$property ) {
+			unset( $property['arg_options'] );
+		}
+		return $schema;
 	}
 
 	public function get_storage_post_type(): string {
@@ -90,6 +386,10 @@ class Subjects_Controller extends WP_REST_Controller {
 			10,
 			2
 		);
+	}
+
+	public function get_endpoint_args_for_subject_schema( $subject_type, $method ): array {
+		return rest_get_endpoint_args_for_schema( $this->schema[ $subject_type ], $method );
 	}
 
 	public function get_item( $request ): WP_REST_Response|WP_Error {
